+++ .github/workflows/matrixflow.yml
@@ -1,0 +1,31 @@
+name: MatrixCI
+
+on: [push, pull_request]
+jobs:
+  build:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python-version: ['3.6', '3.7', '3.8', '3.9', '3.10']
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version  }}
+      - name: Display Python Version
+        run: python --version
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r tests/requirements.txt
+          pip install -e .
+
+      - name: Run Tests
+        run: |
+          export GITHUB_WORKFLOW=1
+          prospector torrentfile
+          prospector tests
+          pytest --cov=torrentfile --cov=tests
