+++ docs/API/index.html
@@ -390,6 +390,21 @@
   
 
     
+      
+      
+      
+
+  
+  
+  
+    <li class="md-nav__item">
+      <a href="../source/" class="md-nav__link">
+        SOURCE
+      </a>
+    </li>
+  
+
+    
   </ul>
 </nav>
                   </div>
@@ -970,33 +985,17 @@ invoked when called from command line, and parses accompanying arguments.</p>
           <span class="mkapi-section-name-body functions">Functions</span> 
         </div>
       <div class="mkapi-section-body functions"><ul class="mkapi-items">
-<li class="function"><code class="mkapi-item-name">create_command</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
-<span class="mkapi-item-type">(torrentfile.MetaFile)</span>
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Execute the create CLI sub-command to create a new torrent metafile.</span></li>
+<li class="function"><code class="mkapi-item-name">execute</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
 
-<li class="function"><code class="mkapi-item-name">edit_command</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
-<span class="mkapi-item-type">(str)</span>
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Execute the edit CLI sub-command with provided arguments.</span></li>
+<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Initialize Command Line Interface for torrentfile.</span></li>
 
-<li class="function"><code class="mkapi-item-name">info_command</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
+<li class="function"><code class="mkapi-item-name">execute</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
 
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Show torrent metafile details to user via stdout.</span></li>
-
-<li class="function"><code class="mkapi-item-name">magnet_command</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">metafile</code><code class="mkapi-object-parenthesis">)</code>
-<span class="mkapi-item-type">(str)</span>
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Create a magnet URI from a Bittorrent meta file.</span></li>
+<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Initialize Command Line Interface for torrentfile.</span></li>
 
 <li class="function"><code class="mkapi-item-name">main</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-parenthesis">)</code>
 
 <span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Initiate main function for CLI script.</span></li>
-
-<li class="function"><code class="mkapi-item-name">main_script</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
-
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Initialize Command Line Interface for torrentfile.</span></li>
-
-<li class="function"><code class="mkapi-item-name">recheck_command</code><code class="mkapi-object-parenthesis">(</code><code class="mkapi-object-signature">args</code><code class="mkapi-object-parenthesis">)</code>
-<span class="mkapi-item-type">(str)</span>
-<span class="mkapi-item-dash">&mdash;</span> <span class="mkapi-item-description">Execute recheck CLI sub-command.</span></li>
 
 
 </ul></div>
