+++ docs/Commands/index.html
@@ -377,6 +377,21 @@
   
 
     
+      
+      
+      
+
+  
+  
+  
+    <li class="md-nav__item">
+      <a href="../source/" class="md-nav__link">
+        SOURCE
+      </a>
+    </li>
+  
+
+    
   </ul>
 </nav>
                   </div>
@@ -477,11 +492,10 @@
 torrentfile [-h] [-i] [-V] [-v]
                 &lt;create&gt; &lt;edit&gt; &lt;magnet&gt; &lt;recheck&gt; ...
 
-CLI utility for editing, creating, checking, and other useful tools for Bittorrent meta files. 
-TorrentFile supports all versions of torrent files.
+CLI Tool for creating, checking, editing... Bittorrent meta files. TorrentFile supports all versions of torrent files.
 
-Optional Arguments
-------------------
+Options
+-------
 -h, --help                          show this help message and exit
 -i, --interactive                   select program options interactively
 -V, --version                       show program version and exit
@@ -497,6 +511,8 @@ Actions
     m (magnet)                        Create magnet url from an existing Bittorrent meta file.
 
     r (recheck, check)                Calculate amount of torrent meta file's content is found on disk.
+
+    i (info)                          Show detailed information about a torrent file.
 </code></pre>
 <hr />
 <h3 id="create">Create</h3>
