+++ docs/LGPLv3/index.html
@@ -384,6 +384,21 @@
   
 
     
+      
+      
+      
+
+  
+  
+  
+    <li class="md-nav__item">
+      <a href="../source/" class="md-nav__link">
+        SOURCE
+      </a>
+    </li>
+  
+
+    
   </ul>
 </nav>
                   </div>
@@ -661,6 +676,21 @@ Library.</p>
           </div>
         </a>
       
+      
+        
+        <a href="../source/" class="md-footer__link md-footer__link--next" aria-label="Next: SOURCE" rel="next">
+          <div class="md-footer__title">
+            <div class="md-ellipsis">
+              <span class="md-footer__direction">
+                Next
+              </span>
+              SOURCE
+            </div>
+          </div>
+          <div class="md-footer__button md-icon">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M4 11v2h12l-5.5 5.5 1.42 1.42L19.84 12l-7.92-7.92L10.5 5.5 16 11H4z"/></svg>
+          </div>
+        </a>
       
     </nav>
