+++ docs/assets/_mkdocstrings.css
@@ -1,0 +1,16 @@
+
+/* Don't capitalize names. */
+h5.doc-heading {
+  text-transform: none !important;
+}
+
+/* Avoid breaking parameters name, etc. in table cells. */
+.doc-contents td code {
+  word-break: normal !important;
+}
+
+/* For pieces of Markdown rendered in table cells. */
+.doc-contents td p {
+  margin-top: 0 !important;
+  margin-bottom: 0 !important;
+}
