+++ docs/changelog/index.html
@@ -539,6 +539,21 @@
   
 
     
+      
+      
+      
+
+  
+  
+  
+    <li class="md-nav__item">
+      <a href="../source/" class="md-nav__link">
+        SOURCE
+      </a>
+    </li>
+  
+
+    
   </ul>
 </nav>
                   </div>
