+++ docs/examples/index.html
@@ -390,6 +390,21 @@
   
 
     
+      
+      
+      
+
+  
+  
+  
+    <li class="md-nav__item">
+      <a href="../source/" class="md-nav__link">
+        SOURCE
+      </a>
+    </li>
+  
+
+    
   </ul>
 </nav>
                   </div>
