+++ docs/sitemap.xml
@@ -2,32 +2,37 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
     <url>
          <loc>https://alexpdev.github.io/torrentfile/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://alexpdev.github.io/torrentfile/API/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://alexpdev.github.io/torrentfile/Commands/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://alexpdev.github.io/torrentfile/LGPLv3/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://alexpdev.github.io/torrentfile/changelog/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
     <url>
          <loc>https://alexpdev.github.io/torrentfile/examples/</loc>
-         <lastmod>2022-02-22</lastmod>
+         <lastmod>2022-02-23</lastmod>
+         <changefreq>daily</changefreq>
+    </url>
+    <url>
+         <loc>https://alexpdev.github.io/torrentfile/source/</loc>
+         <lastmod>2022-02-23</lastmod>
          <changefreq>daily</changefreq>
     </url>
 </urlset>
