+++ site/source.md
@@ -1,0 +1,23 @@
+# TorrentFile
+
+## API and Source
+
+::: torrentfile
+
+::: torrentfile.cli
+
+::: torrentfile.edit
+
+::: torrentfile.hasher
+
+::: torrentfile.interactive
+
+::: torrentfile.recheck
+
+::: torrentfile.torrent
+
+::: torrentfile.utils
+
+::: torrentfile.version
+
+::: tests
