+++ torrentfile/__init__.py
@@ -29,7 +29,7 @@ import logging
 import sys
 
 from torrentfile import interactive, utils
-from torrentfile.cli import magnet_command, main, main_script
+from torrentfile.cli import main, main_script
 from torrentfile.recheck import Checker
 from torrentfile.torrent import TorrentFile, TorrentFileHybrid, TorrentFileV2
 from torrentfile.version import __version__
